LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-offset-keyframes-hidden-subject-expected.txt


FAIL Keyframes with timeline-offsets ignored when timeline is inactive assert_equals: number of frames: Initial keyframes with active view-timeline expected 5 but got 3
